pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tldl"
version = "0.0.2"
description = "A podcast summarizer"
authors = [{ name = "maskedpirate", email = "muratcan.kurtulus@gmail.com" }]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "yt-dlp",
  "streamlit",
  "urllib3",
  "beautifulsoup4",
  "openai-whisper",
  "numpy",
  "langchain-community",
  "langchain-core",
  "TTS",
]

[project.optional-dependencies]
dev = [
  "pytest>=6.0",
  "pytest-cov",
  "black",
  "isort",
  "flake8",
  "python-semantic-release",
]

[project.urls]
Homepage = "https://github.com/makermotion/tldl"
Repository = "https://github.com/makermotion/tldl"

[tool.semantic_release]
version_variables = ["__init__.py:__version__"]
version_toml = ["pyproject.toml:project.version"]
commit_parser = "angular"
major_on_zero = true
tag_format = "v{version}"
commit_message = "{version}\n\nAutomatically generated by python-semantic-release"
build_command = "python -m build"
dist_path = "dist/"
upload_to_pypi = true
upload_to_release = true
branch = "master"

[tool.semantic_release.branches.master]
match = "master"
prerelease = false

[tool.semantic_release.commit_parser_options]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]
major_tags = ["BREAKING CHANGE"]

[tool.semantic_release.changelog]
changelog_file = "CHANGELOG.md"
exclude_commit_patterns = []

[tool.semantic_release.remote]
name = "origin"
type = "github"

[tool.semantic_release.publish]
dist_glob_patterns = ["dist/*"]
upload_to_vcs_release = true
